+++ TDTChocolate/FoundationAdditions/NSArray+TDTFunctionalAdditions.m
@@ -44,6 +44,15 @@
   return [self filteredArrayUsingPredicate:predicate];
 }
 
+- (id)tdt_firstObjectPassingTest:(BOOL (^)(id object))block {
+  for (id object in self) {
+    if (block(object)) {
+      return object;
+    }
+  }
+  return nil;
+}
+
 #pragma mark Partition
 
 - (NSArray *)tdt_partitionUsingBlock:(TDTPredicateBlock)predicate {