+++ CHANGELOG.md
@@ -3,6 +3,8 @@
   * Fix variable name in macro.
   * Add method to check blank string.
 
+  * Add method to extract first object passing test from array.
+
 2.5.1, 22-July-2014
 
   * Fix isSubsetOf to correctly compare mutable/immutable variants.